requirements.txt
argo-ams-library==0.4.2
apimas-django==0.4a3
apimas==0.4a3
specular==0.4a3
Django==1.11.20
django-ckeditor==5.2.2
django-cors-headers==1.1.0
django-rest-swagger==0.3.10
django-reversion==2.0.8
django-suit==0.2.25
djangorestframework==3.4.0
djoser==0.5.2
future==0.15.2
Pillow==4.1.1
beautifulsoup4==4.6.3
